Move CPenaltyShootoutConfiguration traits
